+++ server/src/main/java/org/opensearch/node/Node.java
@@ -1151,6 +1151,13 @@ public class Node implements Closeable {
                 transportService.getTaskManager(),
                 taskCancellationMonitoringSettings
             );
+            final PeerRecoveryTargetService peerRecoveryTargetService = new PeerRecoveryTargetService(
+                threadPool,
+                transportService,
+                recoverySettings,
+                clusterService
+            );
+
             this.nodeService = new NodeService(
                 settings,
                 threadPool,
@@ -1176,7 +1183,8 @@ public class Node implements Closeable {
                 resourceUsageCollectorService,
                 segmentReplicationStatsTracker,
                 repositoryService,
-                admissionControlService
+                admissionControlService,
+                peerRecoveryTargetService
             );
 
             final SearchService searchService = newSearchService(
@@ -1269,8 +1277,7 @@ public class Node implements Closeable {
                     processRecoverySettings(settingsModule.getClusterSettings(), recoverySettings);
                     b.bind(PeerRecoverySourceService.class)
                         .toInstance(new PeerRecoverySourceService(transportService, indicesService, recoverySettings));
-                    b.bind(PeerRecoveryTargetService.class)
-                        .toInstance(new PeerRecoveryTargetService(threadPool, transportService, recoverySettings, clusterService));
+                    b.bind(PeerRecoveryTargetService.class).toInstance(peerRecoveryTargetService);
                     b.bind(SegmentReplicationTargetService.class)
                         .toInstance(
                             new SegmentReplicationTargetService(