+++ test/framework/src/main/java/org/opensearch/test/InternalTestCluster.java
@@ -2736,7 +2736,8 @@ public final class InternalTestCluster extends TestCluster {
                     false,
                     false,
                     false,
-                    false
+                    false,
+                    true
                 );
                 assertThat(
                     "Fielddata size must be 0 on node: " + stats.getNode(),